Use a singleton of monaco-lang-client across multiple editor tabs
